Allow omitting .html in paths (#10)
First step in my quickblog journey. I'd prefer to have the URLs in my
blog stripped from the .html extension. Many web servers do this type
of rewrite already, such as the one used by GitHub Pages where I'm
hosting my blog. But while this works in the hosted version of my blog,
I'm obviously keen to have the same links work when testing locally.

If accepted, the next step would likely be to allow the quickblog
renderer to strip the extension too, if configured to do so. But first
things first :)

Signed-off-by: Anders Eknert <anders@eknert.com>